+++ assets/js/src/integration/index.js
@@ -1,10 +1,15 @@
 import './bricksbuilder';
 import './calderaforms';
 import './contactform7';
+import './elementor';
 import './fluentforms';
 import './formidableforms';
+import './forminator';
 import './gravityforms';
+import './happyforms';
+import './kaliForms';
 import './mc4wp';
+import './newsletter';
 import './ninjaforms';
 import './wpforms';
 import './wsforms';